adjusted anim timing
slowed sprite animation to 12 fps. fixed bug that deformed sprite during movement. was caused by not calculating the position of the second corner of the sprite relative to its position in the game field only at the instant its being drawn, allowing it to change in response to input even after the first corner had been set already, yeilding a deformed sprite in cases of rapid input.