added new object: coin
as a test of the animation system i added a new game object, a coin. this helped to iron out a couple of things that needed generalizing, such as adding a parameter to the game_object constructor to identify an object whose sprite is in perpetual motion like the coin which appears to spin and also refactoring the rendering method to iterate over the collection of game objects stored in the game_animation object. overall it proved the system works well, and provides an excellent segue into the addition of collision detection.