prep for collision
restructured to remove player game_object from game_objects list in preparation for implementing collision detection. haveing it there in the first place was weird anyway, but was only residue of the quick restructure to encapsulate animation properly. made update_sprite method accept a game_object to update, which allows it to be used iteratively. i also cleaned up some repetition there. everything is more or less set up for collision detection as the player is distinct and its position may be considered relative to other game objects.