add rooms wip
added test room class and a room rendering method. room composed of rectangles representing the walls and floor, with the floor being a single rectangle the size of the game window and the walls contained in an array of rectangle objects. pretty straightforward as the rendering method simply changes colors and calls fillRect accessing the rectangle data to draw the walls and floor. this will be replaced by sprites or textures(?) once i get the collision detection and room design worked out.